make mglass::Shape is CRTP-based interface
