+++ src/lib/mglass/CMakeLists.txt
@@ -4,10 +4,13 @@ add_library(mglass STATIC
             "${magnifying-glass_SOURCE_DIR}/include/mglass/primitives.h"
             "${magnifying-glass_SOURCE_DIR}/include/mglass/shape.h"
             "${magnifying-glass_SOURCE_DIR}/include/mglass/shapes.h"
+            "${magnifying-glass_SOURCE_DIR}/include/mglass/ellipse_shape.h"
+            "${magnifying-glass_SOURCE_DIR}/include/mglass/rectangle_shape.h"
             "${magnifying-glass_SOURCE_DIR}/include/mglass/magnifiers.h"
             "mglass.cpp"
             "magnifiers.cpp"
-            "shapes.cpp"
+            "ellipse_shape.cpp"
+            "rectangle_shape.cpp"
             "image.cpp")
 
 enable_extra_compiler_warnings(mglass)
