+++ src/lib/mglass/rectangle_shape.cpp
@@ -1,0 +1,7 @@
+#include "mglass/rectangle_shape.h"
+
+
+namespace mglass::shapes
+{
+
+} // namespace mglass::shapes
